removing more from pixiApp
